Fix issue with processing multiple domains.
Greenlock check function seems to return truthy on the first domain found to be
still fresh. This means adding a new domain to your system will fail because
the code will simply return true and skip the update/register process. To fix this
we manually process one domain at a time.

Also updated to remove hardcoded email address (defaults to a mailinator address
if not provided to avoid breaking old installations) and did an npm update to
upgrade dependencies.